+++ src/main/CharacterSets.ts
@@ -13,6 +13,7 @@ export class CharacterSet {
   POSITIVE: string;
   DOT: string;
   CONTRACTION_20: string;
+  CONTRACTION_30: string;
   useTraditional: boolean;
   useCapital: boolean;
 
@@ -30,15 +31,16 @@ export class CharacterSet {
     this.DOT = Dot;
     this.useTraditional = useTraditional;
     this.useCapital = useCapital;
-    this.CONTRACTION_20 = contraction;
+    this.CONTRACTION_20 = contraction[0];
+    this.CONTRACTION_30 = contraction[1];
   }
 }
 
 const allCharacterSets = [
   //
-  new CharacterSet('零一二三四五六七八九', '十百千万亿兆', '正负', '点', '廿', false, false),
+  new CharacterSet('零一二三四五六七八九', '十百千万亿兆', '正负', '点', '廿卅', false, false),
   new CharacterSet('零壹贰叁肆伍陆柒捌玖', '拾佰仟萬億兆', '正负', '点', '念', false, true),
-  new CharacterSet('零一二三四五六七八九', '十百千萬億兆', '正負', '點', '廿', true, false),
+  new CharacterSet('零一二三四五六七八九', '十百千萬億兆', '正負', '點', '廿卅', true, false),
   new CharacterSet('零壹貳參肆伍陸柒捌玖', '拾佰仟萬億兆', '正負', '點', '念', true, true),
 ];