+++ src/test/Contraction_30.test.ts
@@ -1,0 +1,41 @@
+import {convertNumber, options} from '../main/index';
+import {data} from './data-contraction-30';
+import {generate} from './TestUtils';
+
+const TESTSUITE = 'Contraction_30';
+
+export const option0: options[] = [{use_contraction_30: true}];
+export const option1: options[] = [
+  {use_contraction_30: true, useTraditional: true},
+];
+export const option2: options[] = [
+  {use_contraction_30: true, useCapital: true},
+];
+export const option3: options[] = [
+  {use_contraction_30: true, useTraditional: true, useCapital: true},
+];
+
+test.each(generate(option0, data, 0))(
+  TESTSUITE + '/simplified/normal',
+  (options: options, input: number, expected: string) => {
+    expect(convertNumber(input, options)).toBe(expected);
+  }
+);
+test.each(generate(option1, data, 1))(
+  TESTSUITE + '/traditional/normal',
+  (options: options, input: number, expected: string) => {
+    expect(convertNumber(input, options)).toBe(expected);
+  }
+);
+test.each(generate(option2, data, 2))(
+  TESTSUITE + '/simplified/capital',
+  (options: options, input: number, expected: string) => {
+    expect(convertNumber(input, options)).toBe(expected);
+  }
+);
+test.each(generate(option3, data, 3))(
+  TESTSUITE + '/traditional/capital',
+  (options: options, input: number, expected: string) => {
+    expect(convertNumber(input, options)).toBe(expected);
+  }
+);